pyproject.toml
[tool.poetry]
name = "sbk-scraping"
version = "0.1.0"
description = ""
authors = ["Emmanuel Martínez <emartinez@seebak.com.mx>"]

[[tool.poetry.source]]
name = "pyseebak"
url = "http://pypi.srvc.seebak.com.mx"

[tool.poetry.dependencies]
python = "^3.9"
jmespath = "^0.10.0"
chompjs = "^1.1.6"
pydantic = "^1.8.2"
Scrapy = "^2.5.1"
ipython = "^7.30.1"
environ-config = "^21.2.0"
PyYAML = "^6.0"

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning:.*jmespath*"]

[tool.poetry.dev-dependencies]
pytest = "^6.0"
hypothesis = "^6.30.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
